refactor: sign up page
