+++ src/app/[locale]/login/page.tsx
@@ -5,14 +5,80 @@ import { ROUTES } from '@/sources/routes';
 import { Container, Typography } from '@mui/material';
 import { useTranslations } from 'next-intl';
 import Link from '@mui/material/Link';
+import { useAuthState } from 'react-firebase-hooks/auth';
+import { useEffect, useState } from 'react';
+import { auth, registerWithEmailAndPassword } from '@/firebase';
 
 export default function Login() {
   const t = useTranslations('LoginPage');
+  const [user, loading, error] = useAuthState(auth);
+  const [success, setSuccess] = useState(false);
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [name, setName] = useState('');
+
+  const handleRegister = async () => {
+    try {
+      await registerWithEmailAndPassword(name, email, password);
+      setSuccess(true);
+    } catch (err: unknown) {
+      console.error(err);
+      setSuccess(false);
+    }
+  };
+
+  useEffect(() => {
+    if (loading) return;
+    if (user) {
+      setSuccess(true);
+    }
+    if (error) setSuccess(false);
+  }, [user, loading, error]);
+
   return (
     <Container className="h-screen">
       <Typography variant="h4" component="h2" gutterBottom>
         {t('title')}
       </Typography>
+      <div className="flex flex-col gap-y-4 items-center justify-center min-h-screen bg-gray-100">
+        <input
+          type="text"
+          value={name}
+          onChange={e => setName(e.target.value)}
+          placeholder="Full Name"
+          className="text-black"
+        />
+        <input
+          type="email"
+          value={email}
+          onChange={e => setEmail(e.target.value)}
+          placeholder="E-mail Address"
+          className="text-black"
+        />
+        <input
+          type="password"
+          value={password}
+          onChange={e => setPassword(e.target.value)}
+          placeholder="Password"
+          className="text-black"
+        />
+        <button
+          className="px-6 py-3 text-black rounded-lg shadow-md bg-grey"
+          onClick={handleRegister}
+        >
+          Register
+        </button>
+        {success && (
+          <p className="text-green-600 font-medium">
+            ✅ User successfully registered!
+          </p>
+        )}
+        {error && (
+          <p className="text-red-600 font-medium">
+            ❌ {error.message || 'An error occurred'}
+          </p>
+        )}
+      </div>
       <Typography>{t('noAccount')}</Typography>
       <Link component={IntlLink} href={ROUTES.signup}>
         {t('signup')}
